Reworked RestartFromLocalDevfile test
